refactor: extract council synthesis/review logic into server/councils/synthesis.ts (#291)
Move triggerReview, triggerSynthesis, aggregateSessionResponses, and
finishWithAggregatedSynthesis into a dedicated synthesis module.
Discussion.ts retains orchestration wrappers that inject infrastructure
callbacks (emitLog, broadcastStageChange, watchAutoAdvance) to avoid
circular dependencies.

Also adds *.db.bak-* to .gitignore and removes an accidentally committed
binary database backup file.
